% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_BE_aggregate_fcs_files.R
\name{BE_aggregate_fcs_files}
\alias{BE_aggregate_fcs_files}
\title{BE_aggregate_fcs_files}
\usage{
BE_aggregate_fcs_files(patient_names, fsom, fcs_dir, markers,
  metadata_patients)
}
\arguments{
\item{patient_names}{vector containing the fcs names of the patients, and their short names as names}

\item{fsom}{FlowSOM map to match the fcs files to}

\item{fcs_dir}{Directory of the FCS files}

\item{markers}{markers to keep in the final table}

\item{metadata_patients}{metadata information about the patients. Should contain at least one column "DATEOFCYTOFEXPERIMENT", and have the short names of the patients as rownames.}
}
\value{
a data table containing the expression of the markers of interest in the cells of all patients, with three extracolumns : cluster_id, patient_id and day_id.
}
\description{
BE_aggregate_fcs_files
}
\examples{

aggreg_table <- BE_aggregate_fcs_files(patient_names = recip_names, fsom = fsom_recip,
fcs_dir = fcs_dir, markers = functional_marks,
metadata_patients = samp_recip)
}
