% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_BE_compute_MFI_per_metaclust.R
\name{BE_compute_MFI_per_metaclust}
\alias{BE_compute_MFI_per_metaclust}
\title{Title}
\usage{
BE_compute_MFI_per_metaclust(aggreg_table, fsom)
}
\arguments{
\item{aggreg_table}{Table containing all cells from patients of interest, as returned by the function BE_aggregate_to_table}

\item{fsom}{FlowSOM object}
}
\value{
A dataframe of n(patients)*[p(markers)*m(metaclusters)], containing the mean expression of each functional marker in each metacluster per patient
}
\description{
Title
}
\examples{

aggreg_table <- BE_aggregate_fcs_files(patient_names = recip_names[1:2], fsom = fsom_recip,
fcs_dir = fcs_dir, markers = functional_marks,
metadata_patients = samp_recip)

mean_MFI_df <- BE_compute_MFI_per_metaclust(aggreg_table, fsom)
}
