% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_BE_QC.R
\name{BE_QC}
\alias{BE_QC}
\title{BE_QC}
\usage{
BE_QC(aggreg_table, png_name_1, aggreg_table_rescaled, png_name_2)
}
\arguments{
\item{aggreg_table}{Table containing all cells from patients of interest, as returned by the function BE_aggregate_to_table}

\item{png_name_1}{Name of the output file correspnding to the original values}

\item{aggreg_table_rescaled}{Table aggreg_table, but with re-scaled values}

\item{png_name_2}{Name of the output file correspnding to the re-scaled values}
}
\value{
Two png files
}
\description{
BE_QC
}
\examples{

aggreg_table <- BE_aggregate_fcs_files(patient_names = recip_names[1:2], fsom = fsom_recip,
fcs_dir = fcs_dir, markers = functional_marks,
metadata_patients = samp_recip)
BE_QC(aggreg_table, "my_output_file.png")
}
