% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_compatibility_scores.R
\name{compute_compatibility_score}
\alias{compute_compatibility_score}
\title{compute_compatibility_scores}
\usage{
compute_compatibility_score(samp_rd)
}
\arguments{
\item{samp_rd}{metadata table containing the information on donors and recipients}
}
\value{
the differences between donors and recipients
}
\description{
compute_compatibility_scores
}
\examples{

compatibility_scores <- compute_compatibility_score(samp_rd)
}
