% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_marker_MFIs_per_metacluster.R
\name{compute_marker_MFIs_per_metacluster}
\alias{compute_marker_MFIs_per_metacluster}
\title{Title}
\usage{
compute_marker_MFIs_per_metacluster(ff_agg, pctgs_meta, fsom, functional_marks)
}
\arguments{
\item{ff_agg}{aggregated cells from each patient, as returned by AggregateFlowFrames FlowSOM function}

\item{pctgs_meta}{Matrix of n(patients) * p(metaclusters) containing the cell percentages from patients in fsom metaclusters}

\item{fsom}{FlowSOM object}

\item{functional_marks}{list of markers to use to compute median MFIs in fsom metaclusters}
}
\value{
description of output
}
\description{
Title
}
