% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_compute_pca_distance_scores.R
\name{compute_pca_distance_scores}
\alias{compute_pca_distance_scores}
\title{compute_pca_distance_scores}
\usage{
compute_pca_distance_scores(pca_metaclust, samp_rd)
}
\arguments{
\item{pca_metaclust}{pca on fsom metaclusters, result of the prcomp function}

\item{samp_rd}{metadata on donors and recipients}
}
\value{
distances between donors and recipients on the pca's 2 first principal components. A score of one corresponds to the smallest distance ( = maximum resemblance), a score of 0.5 corresponds to the biggest distance.
}
\description{
compute_pca_distance_scores
}
\examples{

compute_pca_distance_scores(pca_metaclust, samp_rd)

}
