% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_compute_pctg_scores.R
\name{compute_pctg_scores}
\alias{compute_pctg_scores}
\title{compute_pctg_scores}
\usage{
compute_pctg_scores(pctgs_meta, samp_rd)
}
\arguments{
\item{pctgs_meta}{pca on fsom metaclusters, result of the prcomp function}

\item{samp_rd}{metadata on donors and recipients}
}
\value{
distances between donors and recipients on the pca's 2 first principal components. A score of one corresponds to the smallest distance ( = maximum resemblance), a score of 0.5 corresponds to the biggest distance.
}
\description{
compute_pctg_scores
}
\examples{

compute_pctg_scores(pctgs_meta, samp_rd)

}
