% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_DEstats_and_graphinfo.R
\name{export_DEstats_and_graphinfo}
\alias{export_DEstats_and_graphinfo}
\title{export_DEstats_and_graphinfo}
\usage{
export_DEstats_and_graphinfo(
  gr,
  common_features,
  common_beh_features,
  norm_data,
  comparison,
  folder_path,
  common_ft_name,
  common_beh_ft_name
)
}
\arguments{
\item{gr}{A graph as returned by the function plot_correlations}

\item{common_features}{The features that are common to both cohorts}

\item{common_beh_features}{The features that have the same behaviour in both cohorts}

\item{norm_data}{The normalised data}

\item{comparison}{The comparison to perform: either "TNT" (non tolerant vs tolerant) or "PTST" (primary vs secondary tolerant)}

\item{folder_path}{The path to the folder where files should be exported}

\item{common_ft_name}{The name of the file that will contain information on the common features}

\item{common_beh_ft_name}{The name of the file that will contain information on the common beh features}
}
\value{
excel files
}
\description{
export_DEstats_and_graphinfo
}
