% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cyto_extract_patients_funct_marks_per_metaclus.R
\name{extract_funct_markers}
\alias{extract_funct_markers}
\title{extract_funct_markers}
\usage{
extract_funct_markers(
  recip_names,
  fsom,
  funct_marks,
  min_ref,
  max_ref,
  files2rescale,
  samp_patients
)
}
\arguments{
\item{recip_names}{vector containing the names of recipients' fcs files}

\item{fsom}{FlowSOM object}

\item{funct_marks}{functional markers for which to extract MFIs per metacluster per patient}

\item{min_ref}{vector of minimal values to use for rescaling}

\item{max_ref}{vector of maximal values to use for rescaling}

\item{files2rescale}{vector of names of the files that need to be rescaled}

\item{samp_patients}{table providing information on the patients (group, batch, ...)}
}
\value{
a list of MFIs matrices for each column to use
}
\description{
Generates a list of matrices (one matrix per metacluster), containing the
expression of the functional markers in the cells from that cluster, from
all patients
}
\examples{
funct_matrices <- extract_funct_markers(recip_names, fsom, fcs_dir, funct_marks,
                                        min_ref, max_ref, files2rescale = c("D1073"))

}
