% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabo_merge_metabo_metadata.R
\name{extract_info_metabo}
\alias{extract_info_metabo}
\title{extract_info_metabo}
\usage{
extract_info_metabo(metadata_file, metabo_file)
}
\arguments{
\item{metadata_file}{dataframe containing patients in rows and some information in columns}

\item{metabo_file}{excel file containing the metabolomics data}
}
\value{
A list, containing patient metadata, the metabolites data and the pathways and subpathways of these metabolites.
}
\description{
extract_info_metabo
}
\examples{

merged <- merge_metabo_metadata(metadata_file = samp_recip,
 metabo_file = "~/Documents/VIB/Projects/Integrative_Paris/documents_22:02:18/CYTOF_David_Michonneau/Metabolomic local cohort Saint-Louis_filtered.xlsx")
}
