% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_common_behaviour_features.R
\name{find_common_behaviour_features}
\alias{find_common_behaviour_features}
\title{find_common_behaviour_features}
\usage{
find_common_behaviour_features(list_common_features, file_path,
  common_ft_file_name, common_beh_file_name)
}
\arguments{
\item{list_common_features}{The list containing the common features and the local and other cohort dataframes, as returned by the function find_common_features}

\item{file_path}{The path to the folder where all results should be stored}

\item{common_ft_file_name}{the excel file name to save common features}

\item{common_beh_file_name}{the excel file name to save common features with the same behaviour}
}
\value{
The features that had the same behaviour in both cohorts
}
\description{
Function to find features that had a similar group behaviour in the two cohorts
}
