% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_common_features.R
\name{find_common_features}
\alias{find_common_features}
\title{Find_common_features}
\usage{
find_common_features(selected_features, ndata, other_features_file,
  other_norm_data_file, file_path, file_name)
}
\arguments{
\item{selected_features}{The selected features in this cohort}

\item{ndata}{The normalised data of this cohort (with patients in rows)}

\item{other_features_file}{The path to the excel file containing the selected features of the other cohort}

\item{other_norm_data_file}{The path to the RDS file containing the normalised data of the other cohort (with patients in rows)}

\item{file_path}{Where new results should be saved}

\item{file_name}{The file name}
}
\value{
A list containing 1) the common features between the 2 cohorts
2) the dataframe of the common features in the local cohort
3) the dataframe of the common features in the other cohort
}
\description{
Find_common_features
}
