% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabo_find_subpathways_driving_PC.R
\name{find_subpatways_driving_PC}
\alias{find_subpatways_driving_PC}
\title{find_subpatways_driving_PC}
\usage{
find_subpatways_driving_PC(meta_metabo, pca, PC)
}
\arguments{
\item{meta_metabo}{A dataframe containing the metabolites in columns, and 3 rows: the sub-pathway, the superpathway, and the metabolite name}

\item{pca}{The result of a PCA, as returned by the ade4::dudi.pca function}

\item{PC}{A numeric, corresponding to the principal component to be investigated}
}
\value{
A list containing: 1) the "ordered_subpathways", a vector containing all subpathways ordered by importance along the PC of interect
and 2) "princ_axes", a dataframe containing both the meta_metabo data and the importance of the different metabolites along each PC of the pca
}
\description{
Identifies which which sub-pathways drive a pca axes of interest
}
\examples{
pca <- ade4::dudi.pca(metabo_data)
result_pc1 <- find_subpatways_driving_PC(meta_metabo = meta_metabo, pca = pca, PC = 1)
}
