% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_fsom2fsom_meta.R
\name{fsom2fsom_meta}
\alias{fsom2fsom_meta}
\title{fsom2fsom_meta}
\usage{
fsom2fsom_meta(fsom, colsToUse, pctgs_patients, plot_size)
}
\arguments{
\item{fsom}{FlowSOM object}

\item{colsToUse}{names of the fluorochromes of the markers to use}

\item{pctgs_patients}{pctgs of cells of each patient in fsom clusters, matrix of n(patients) * p(clusters)}

\item{plot_size}{Parameter to set the size of the nodes in fsom map. Can be set to "equal_size", or the node size can be proportional to the number of cells from one patient. In this case, plot_size should be a numeric, corresponding to the desired patient.}
}
\value{
New FlowSOM object, with metaclusters instead of clusters.
}
\description{
This fuction reconstructs a FlowSOM object structure with metaclusters instead of clusters. The resulting FlowSOM object can be plotted with functions from the FlowSOM package such as PlotStars, PlotNumbers...
}
\examples{
fsom_meta_recip <- fsom2fsom_meta(fsom = fsom_recip, colsToUse = markersToPlot,
pctgs_patients = pctgs_recip, plot_size = "equal_size")
PlotNumbers(fsom_meta_recip$FlowSOM)


}
