% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integ_gather_DRDR.R
\name{gather_DRDR}
\alias{gather_DRDR}
\title{Gather Donors, Recipients and D-R}
\usage{
gather_DRDR(
  data_couples,
  beh_couples = NULL,
  data_recip,
  beh_recip = NULL,
  data_donors,
  beh_donors = NULL,
  couple_info,
  couple_info_2keep
)
}
\arguments{
\item{data_couples}{The path to the matrix with couples in rows and selected feaatures in columns}

\item{beh_couples}{The path to the list of couple features that had the same behaviour in the 2 cohorts}

\item{data_recip}{The path to the matrix with recip in rows and selected feaatures in columns}

\item{beh_recip}{The path to the list of recip features that had the same behaviour in the 2 cohorts}

\item{data_donors}{The path to the matrix with donors in rows and selected feaatures in columns}

\item{beh_donors}{The path to the list of donor features that had the same behaviour in the 2 cohorts}

\item{couple_info}{The dataframe containing all donors and recipients in rows and meta information on them}

\item{couple_info_2keep}{The columns in couple info that you wish to keep}
}
\value{
a dataframe containing the groupes features of D, R, D-R and some meta info
}
\description{
This function takes as imput: information on the patients under the form of
a couple_info dataframe, as well as the path to the matrices of the selected
features for recipients, donors and couples.
optionally, it can also take the lists of genes that had the same behaviour
in the two cohorts and select only these features.
It returns a big dataframe, regrouping the selected features of recipients,
donors and couples, as well as some meta info.
}
