% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_counts.R
\name{generate_counts}
\alias{generate_counts}
\title{generate_counts}
\usage{
generate_counts(recip_names, fsom, pdf_name, fcs_dir)
}
\arguments{
\item{recip_names}{vector containing the names of recipients' fcs files}

\item{fsom}{FlowSOM object}

\item{pdf_name}{choose a name for the exported pdf which will contain one fsom map per patient}

\item{fcs_dir}{path to the directory containing fcs files, corresponding to recip_names}
}
\value{
a matrix and pdf
}
\description{
Generates a matrix of counts and pdf fsom files of all cells mapped to fsom for each patient
}
\examples{
bla
}
