% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_generate_donor_pctgs.R
\name{generate_donor_pctgs}
\alias{generate_donor_pctgs}
\title{generate_donor_pctgs}
\usage{
generate_donor_pctgs(donor_names, fsom, pdf_name, fcs_dir, output_dir)
}
\arguments{
\item{donor_names}{vector containing the names of donors' fcs files}

\item{fsom}{FlowSOM object}

\item{pdf_name}{choose a name for the exported pdf which will contain one fsom map per patient}

\item{fcs_dir}{path to the directory containing fcs files, corresponding to donor_names}

\item{output_dir}{}
}
\value{
a pctgs matrix and pdf containing one map per patient
}
\description{
Generates a matrix of pctgs and pdf fsom files of pctgs of each patient's cells mapped to fsom clusters
}
\examples{
pctgs <- generate_donor_pctgs(donor_names, fsom, pdf_name = "my_pdf.pdf", fcs_dir)
}
