% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_generate_meta_MFIs.R
\name{generate_meta_MFIs}
\alias{generate_meta_MFIs}
\title{generate_meta_MFIs}
\usage{
generate_meta_MFIs(recip_names, fsom, cols_to_use, min_ref, max_ref,
  files2rescale)
}
\arguments{
\item{recip_names}{vector containing the names of recipients' fcs files}

\item{fsom}{FlowSOM object}

\item{cols_to_use}{functional markers for which to extract MFIs per metacluster per patient}

\item{min_ref}{vector of minimal values to use for rescaling}

\item{max_ref}{vector of maximal values to use for rescaling}

\item{files2rescale}{vector of names of the files that need to be rescaled}
}
\value{
a list of MFIs matrices for each column to use
}
\description{
Generates a matrix of MFIs of each patient's cells mapped to fsom metaclusters
}
\examples{
pctgs <- generate_pctgs(recip_names, fsom, pdf_name = "my_pdf.pdf", fcs_dir)
}
