% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_compute_MFIs_per_metacluster.R,
%   R/cyto_generate_pctgs.R
\name{generate_pctgs}
\alias{generate_pctgs}
\alias{generate_pctgs}
\title{generate_pctgs}
\usage{
generate_pctgs(recip_names, fsom, pdf_name, fcs_dir, output_dir)

generate_pctgs(recip_names, fsom, pdf_name, fcs_dir, output_dir)
}
\arguments{
\item{recip_names}{vector containing the names of recipients' fcs files}

\item{fsom}{FlowSOM object}

\item{pdf_name}{choose a name for the exported pdf which will contain one fsom map per patient}

\item{fcs_dir}{path to the directory containing fcs files, corresponding to recip_names}

\item{recip_names}{vector containing the names of recipients' fcs files}

\item{fsom}{FlowSOM object}

\item{pdf_name}{choose a name for the exported pdf which will contain one fsom map per patient}

\item{fcs_dir}{path to the directory containing fcs files, corresponding to recip_names}
}
\value{
a pctgs matrix and pdf containing one map per patient

a pctgs matrix and pdf containing one map per patient
}
\description{
Generates a matrix of pctgs and pdf fsom files of pctgs of each patient's cells mapped to fsom clusters

Generates a matrix of pctgs and pdf fsom files of pctgs of each patient's cells mapped to fsom clusters
}
\examples{
pctgs <- generate_pctgs(recip_names, fsom, pdf_name = "my_pdf.pdf", fcs_dir)
pctgs <- generate_pctgs(recip_names, fsom, pdf_name = "my_pdf.pdf", fcs_dir)
}
