% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_generate_pctgs_cryostem.R
\name{generate_pctgs_cryostem}
\alias{generate_pctgs_cryostem}
\title{generate_pctgs_cryostem}
\usage{
generate_pctgs_cryostem(recip_names, fsom, pdf_name, fcs_dir, output_dir,
  min_ref = NULL, max_ref = NULL, files2rescale = NULL)
}
\arguments{
\item{recip_names}{vector containing the names of recipients' fcs files}

\item{fsom}{FlowSOM object}

\item{pdf_name}{choose a name for the exported pdf which will contain one fsom map per patient}

\item{fcs_dir}{path to the directory containing fcs files, corresponding to recip_names}

\item{min_ref}{vector of minimum values to use for rescaling}

\item{max_ref}{vector of maximum values to use for rescaling}

\item{files2rescale}{vector of names of the files to rescale}
}
\value{
a pctgs matrix and pdf containing one map per patient + information on the % of cells that were too far from cluster center
}
\description{
Generates a matrix of pctgs and pdf fsom files of pctgs of each patient's cells mapped to fsom clusters
! doesn't use CD19 to map the cells, since CD19 is unusable in the cryostem cohort
}
\examples{
pctgs <- generate_pctgs(recip_names, fsom, pdf_name = "my_pdf.pdf", fcs_dir)
}
