% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_analysis_results.R
\name{ggplot_analysis_results}
\alias{ggplot_analysis_results}
\title{ggplot_analysis_results}
\usage{
ggplot_analysis_results(analysis, perplexity_nb = 8, data_matrix,
  metadata, col_by, shape_by)
}
\arguments{
\item{analysis}{Type of analysis to be performed. For now, either "PCA" or "tSNE" can be run.}

\item{perplexity_nb}{If tSNE analysis, choose the perplexity parameter. Default is 8.}

\item{data_matrix}{Matrix on which to run the PCA or tSNE analysis, with patients in rows and parameters in columns.}

\item{metadata}{Dataframe containing information about the patients, with patients in rows and variables in columns.}

\item{col_by}{Metadata column to use as a factor to color points in analysis.}

\item{shape_by}{Metadata column to use as a factor to display the shape of points in analysis.}
}
\value{
ggplot
}
\description{
ggplot_analysis_results
}
\examples{
ggplot_analysis_results("tSNE", perplexity = 8, data_matrix = pctgs_recip, metadata = samp_recip,
col_by = "DATEOFCYTOFEXPERIMENT", shape_by = "GROUP")


}
