% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_identify_fsom_cellPopulations.R
\name{identify_fsom_cellPopulations}
\alias{identify_fsom_cellPopulations}
\title{identify_fsom_cellPopulations}
\usage{
identify_fsom_cellPopulations(fsom, prettyMarkerNames, cellTypes, pdf_name,
  view)
}
\arguments{
\item{fsom}{A FlowSOM object}

\item{prettyMarkerNames}{vector containing the biological names of the markers and the associated fluorochromes}

\item{cellTypes}{A list of the markers expressed in each cell type, as returned by the "markers_of_cellTypes" function}

\item{pdf_name}{Name of the pdf to be generated, which will contain one map per cell type, with the clusters corresponding to the cell type highlighted in red.}

\item{view}{either "grid" or "MST", given as a parameter to FlowSOM::PlotStars}
}
\value{
Cell types associated to the clusters in the FlowSOM object
}
\description{
Matches cell types to the clusters of a FlowSOM object.
}
\examples{

celllabels <- identify_fsom_cellPopulations(fsom = fsom_meta_recip, prettyMarkerNames, cellTypes,
pdf_name = "pops_filt_marks_grid.pdf", view="MST")
}
