% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_patient_info.R
\name{import_patient_info}
\alias{import_patient_info}
\title{import_patient_info}
\usage{
import_patient_info(data_synthesis_file, patient_names, patient_type)
}
\arguments{
\item{data_synthesis_file}{The excel file name}

\item{patient_names}{The FCS names of the patients for whom information should be kept}
}
\value{
A data frame corresponding to the excel file, with only the patients of interest and no NAs
}
\description{
Imports an excel table containing all the metadata concerning patients
Columns conataining NAs are removed
Columns containing characters are turned into factors
Columns containing dates are imported as such
}
\examples{

mytable <- import_patient_info(data_synthesis_file = "~/Documents/VIB/Projects/Integrative_Paris/documents_22:02:18/CYTOF_David_Michonneau/Data synthesis local cohort Saint-Louis 032018_modified.xlsx",
patient_names = recip_names)
}
