% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_of_cellTypes.R
\name{markers_of_cellTypes}
\alias{markers_of_cellTypes}
\title{markers_of_cellTypes}
\usage{
markers_of_cellTypes(marker_table)
}
\arguments{
\item{marker_table}{A dataframe containing the populations in rows and the markers in columns. Values are either "+","-" or "NA".}
}
\value{
A list of markers describing each cell population.
}
\description{
Extracts the necessary information from a table describing which markers are expressed or not in each cell population.
Returns a list of all populations, with the markers that allow to identify them.
}
\examples{

cellTypes <- markers_of_cellTypes(marker_table = pop_mark)
}
