% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabo_plot_cor_2_cohorts.R
\name{metabo_cor_2_cohorts}
\alias{metabo_cor_2_cohorts}
\title{metabo_ plot correlation between 2 cohorts}
\usage{
metabo_cor_2_cohorts(model_stat1, norm_data1, model_stat2, norm_data2,
  stat_of_interest, perm_distribution = FALSE, meta_metabo)
}
\arguments{
\item{model_stat1}{The model stat of the cryostem cohort}

\item{norm_data1}{The norm_data of the cryostem cohort}

\item{model_stat2}{The model stat of the st louis cohort}

\item{norm_data2}{The norm_data of the st louis cohort}

\item{stat_of_interest}{2 = global AUC, 3 = tol 1 vs non, 4 = tol 2 vs non, 5 = tol 1 vs tol 2}

\item{perm_distribution}{Boolean defining if the stats to be compared between the two cohorts are AUCs or permutation quantiles}

\item{meta_metabo}{the information on the pathways and subpathways}
}
\value{
plots
}
\description{
metabo_ plot correlation between 2 cohorts
}
