% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabo_preprocess.R
\name{metabo_preprocess}
\alias{metabo_preprocess}
\title{metabo_preprocess}
\usage{
metabo_preprocess(patient_type, data_metabolites, meta_metabo,
  other_cohort_metabolites, rd_meta, names_patients, pdf_name,
  pdf_variance_name, save_results, save_res_repository)
}
\arguments{
\item{patient_type}{A charcter string, either "donors", "recipients" or "donors and recipients"}

\item{data_metabolites}{The metabo data frame, with patients in rows and metabolites in columns}

\item{meta_metabo}{A dataframe containing information on the metabolites, as returned by the extract_info_metabo function}

\item{other_cohort_metabolites}{The names of the metabolites from the unused cohort}

\item{rd_meta}{The dataframe containing clinical information, as returned by the extract_info_metabo function}

\item{names_patients}{A character string containing the names of the patients to analyse}

\item{pdf_name}{Name of the figure containing the metabolites with too many NAs}

\item{pdf_variance_name}{Name of the figure containing the metabolite variance, and the cutoff applied to filter lowly variable metabolites}

\item{save_results}{If True, the folder where the results should be saved has to be provided to the "save_res_repository"}

\item{save_res_repository}{The folder where the results should be saved}
}
\value{
a list of objects : the log data, normalised data, metabo information, and big_matrix containing the metabo + clinical info
}
\description{
metabo_preprocess
}
