% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations_value.R
\name{perm_val}
\alias{perm_val}
\title{Permutations value}
\usage{
perm_val(data, nb_perm, ind_feature)
}
\arguments{
\item{data}{A normalised dataset (dataframe with patients in rows and features in columns)}

\item{nb_perm}{The number of permutations to generate the permutation distribution}

\item{ind_feature}{the feature index, a column number}
}
\value{
the quantile value of the feature in the permutation distribution
}
\description{
This function takes as input a dataset and a feature id and returns the quantile value of the prediction score of the feature compared to the permutation distribution
}
