% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_values_with_demographics.R
\name{perm_val_demo}
\alias{perm_val_demo}
\title{Permutations value with demographics}
\usage{
perm_val_demo(data, nb_perm, ind_feature, orig_AUC)
}
\arguments{
\item{data}{A normalised dataset (dataframe with patients in rows and features in columns)}

\item{nb_perm}{The number of permutations to generate the permutation distribution}

\item{ind_feature}{the feature index, a column number}

\item{orig_AUC}{the original AUC, computed only on the demographics}
}
\value{
the quantile value of the feature in the permutation distribution
}
\description{
This function takes as input a dataset and a feature id and returns the quantile value of the prediction score of the feature compared to the permutation distribution
}
