% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_values_log_reg_demo.R
\name{perm_val_LR_demo}
\alias{perm_val_LR_demo}
\title{Permutations value log regression with demographics}
\usage{
perm_val_LR_demo(data, nb_perm, ind_feature)
}
\arguments{
\item{data}{A normalised dataset (dataframe with patients in rows and features in columns)}

\item{nb_perm}{The number of permutations to generate the permutation distribution}

\item{ind_feature}{the feature index, a column number}
}
\value{
the quantile value of the feature in the permutation distribution
}
\description{
This function takes as input a dataset and a feature id and returns
1) the quantile value of the prediction score of the feature compared to
the permutation distribution
2) the quantile value when taking the age and gender demographics into account
}
