% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate_markers.R
\name{plot_aggregate_markers}
\alias{plot_aggregate_markers}
\title{cyto_plot_aggregate_markers}
\usage{
plot_aggregate_markers(samp_recip, arrange_by, prettyMarkerNames, pheno_marks,
  png_name, ff_agg)
}
\arguments{
\item{samp_recip}{Metadata about the patients}

\item{prettyMarkerNames}{A vector containing the biological marker names as well as the fluorochrome names}

\item{pheno_marks}{A vector containing phenotypic markers}

\item{png_name}{Name of the png file to be exported}

\item{ff_agg}{Aggregated cells from each patient, as returned by the AggregateFlowFrames FlowSOM function}
}
\value{
A png
}
\description{
Plots expression of phenotypic markers across patients (QC):
}
\examples{
bla
}
