% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{plot_correlations}
\usage{
plot_correlations(ft_df, compar)
}
\arguments{
\item{ft_df}{The dataframe with samples in rows, features + group in columns}

\item{compar}{The comparison: either "TNT" or "PS"}
}
\value{
A plot
}
\description{
plot_correlations
}
