% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations2.R
\name{plot_correlations2}
\alias{plot_correlations2}
\title{plot_correlations}
\usage{
plot_correlations2(
  ft_df,
  ft_file_other_cohort,
  compar = "TNT",
  node_color = NULL,
  pval_threshold
)
}
\arguments{
\item{ft_df}{The dataframe with samples in rows, features + group in columns}

\item{compar}{The comparison: either "TNT" or "PS"}
}
\value{
A plot and the graph
}
\description{
plot_correlations
}
