% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_dr.R
\name{plot_density_dr}
\alias{plot_density_dr}
\title{plot_density_dr}
\usage{
plot_density_dr(samp_rd_tmp, selected_group, featur, mat_pheno_funct)
}
\arguments{
\item{samp_rd_tmp}{the clinical data table, with rownames = rownames of mat_pheno_funct}

\item{selected_group}{the group of choice ("non_tolerant", "primary_tolerant" or "secondary_tolerant")}

\item{featur}{the feture to plot, must be in colnames of mat_pheno_funct}

\item{mat_pheno_funct}{the expression data matrix, must have same rownames as samp_rd_tmp}
}
\value{
a ggplot
}
\description{
A function to plot both densities of donors and recipients in the 3 groups,
to help compare differences in expression
}
