% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_patients_funct_marks_per_petaclus.R
\name{plot_funct_markers}
\alias{plot_funct_markers}
\title{plot_funct_markers}
\usage{
plot_funct_markers(metaclust_values, patient_names, samp_patients,
  png_name)
}
\arguments{
\item{metaclust_values}{list of matrices containing the cell expression of
the functional markers for all patients, per metacluster, as generated by the
function extract_funct_markers}

\item{patient_names}{list of the patient names, same as was passed to the
extract_funt_markers function}
}
\value{
a list of MFIs matrices for each column to use
}
\description{
Plot the expression of functional markers in all patients' cells, per
metacluster. This information can be generated with the
extract_funct_markers() function.
}
\examples{
funct_matrices <- extract_funct_markers(recip_names, fsom, fcs_dir, cols_to_use,
                                        min_ref, max_ref, files2rescale = c("D1073"))

}
