% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie.R
\name{plot_pie}
\alias{plot_pie}
\title{plot_pie}
\usage{
plot_pie(df, colors)
}
\arguments{
\item{df}{A dataframe, with 2 columns: "group", containing the names of the groups, and "value", containing the associated percentages}

\item{colors}{The colors to use for every group in the pie}
}
\value{
Plots the pie
}
\description{
This fuction plots a pie
}
\examples{
df <- data.frame(
group = c("Male", "Female", "Child"),
value = c(25, 25, 50))

plot_pie(df, colors = c("blue","green","yellow"))



}
