% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_gender_age.R
\name{plots_gender_age}
\alias{plots_gender_age}
\title{plots_gender_age}
\usage{
plots_gender_age(norm_data, features, pdf_name, compar)
}
\arguments{
\item{norm_data}{A dataframe containing patients in rows, features in columns, and the columns "group", "gender_comp" and "age_recip"}

\item{features}{A vector of the features of interest}

\item{pdf_name}{The pdf name}
}
\value{
a PDF
}
\description{
Generates forest plots and other informative plots to better understand the
role of a feature in tolerance with respect to age and gender
}
