% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf_tol1_tol2.R
\name{rf_tol1_tol2}
\alias{rf_tol1_tol2}
\title{Title}
\usage{
rf_tol1_tol2(df_orig, mtry = 3, ntree = 10000,
  package_2use = "randomForest")
}
\arguments{
\item{df_orig}{Original dataframe on which rf should be run. Should contain a "group" column, with "primary_tolerant", "secondary_tolerant" or "non_tolerant" factor levels}

\item{mtry}{Number of parameters to be tested by RF at every split}

\item{ntree}{Number of trees in the RF}

\item{package_2use}{The random forest implementation to use (either "randomForest" or "ranger")}
}
\value{
RF results on classification of tol vs non-tol patients
}
\description{
Title
}
