% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf_tol_non_tol.R
\name{rf_tol_non_tol}
\alias{rf_tol_non_tol}
\title{rf_tol_non_tol}
\usage{
rf_tol_non_tol(df_orig, mtry = 3, ntree = 10000)
}
\arguments{
\item{df_orig}{Original dataframe on which rf should be run. Should contain a "group" column, with "primary_tolerant", "secondary_tolerant" or "non_tolerant" factor levels}

\item{mtry}{Number of parameters to be tested by RF at every split}

\item{ntree}{Number of trees in the RF}
}
\value{
RF results on classification of tol vs non-tol patients
}
\description{
rf_tol_non_tol
}
