% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_save_genes_annot.R
\name{save_genes_annot}
\alias{save_genes_annot}
\title{Save_genes_annot}
\usage{
save_genes_annot(norm_data, path_obj, selected_genes, compar_design,
  graph = NULL, path_annot_obj)
}
\arguments{
\item{norm_data}{The normalised data matrix}

\item{path_obj}{Where to save the vector of selected genes}

\item{selected_genes}{The selected genes to save}

\item{compar_design}{Either "TNT" or "PS"}

\item{graph}{The graph from which correlation information needs to be saved}

\item{path_annot_obj}{Where to save the annotation object}
}
\value{
Two excel objects: the vector of selected genes and the dataframe of
the annotated selected genes.
}
\description{
Function to save the names of the selected genes, as well as a data frame 
containing the FC and pvalues for these genes, and the graph informations
(ie how the genes were correlated)
}
