% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_features.R
\name{select_features}
\alias{select_features}
\title{Select_features}
\usage{
select_features(
  perm_vals,
  norm_data,
  alone_or_demo = 1,
  threshold,
  file_path,
  file_name
)
}
\arguments{
\item{perm_vals}{The permutation distributions results}

\item{norm_data}{The normalised data with patients in rows and features in columns}

\item{alone_or_demo}{1: the permutation values for genes alone, 2: the permutation values for models incorporating the genes + age + gender}

\item{threshold}{The threshold to use}

\item{file_path}{The file_path}

\item{file_name}{The file_name (typically: perm_val_TNT_95_thresh.xlsx)}
}
\value{
A vector containing the selected genes
}
\description{
Select features that had a value above a cetain threshold in the permutaion
values results.
}
