% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rf_tree.R
\name{tree_func}
\alias{tree_func}
\title{tree_func}
\usage{
tree_func(final_model)
}
\arguments{
\item{final_model}{result of a random forest}
}
\value{
plots a tree
}
\description{
tree_func
}
\examples{
tree_func(rf_result)
}
