% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_features_in_boxplots.R
\name{visualise_features_in_boxplots}
\alias{visualise_features_in_boxplots}
\title{visualise_features_in_boxplots}
\usage{
visualise_features_in_boxplots(list_common_features)
}
\arguments{
\item{list_common_features}{The list containing the common features and the local and other cohort dataframes, as returned by the function find_common_features}
}
\value{
Boxplots
}
\description{
visualise_features_in_boxplots
}
