% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_fcs_to_agg.R
\name{fcs_to_agg}
\alias{fcs_to_agg}
\title{Title}
\usage{
fcs_to_agg(fcs_dir, fcs_names, seed, cTotal, output_name)
}
\arguments{
\item{fcs_dir}{param}

\item{fcs_names}{param}

\item{seed}{param}

\item{cTotal}{param}

\item{output_name}{param}
}
\value{
aggregated flowframe as returned by the FlowSOM::AggregateFlowFrames function
}
\description{
Title
}
\examples{
bla
}
